pyproject.toml
[project]
name = "linebot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "wwwxxch", email = "chihhui.x@gmail.com" }
]
dependencies = [
    "line-bot-sdk>=3.14.2",
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/linebot"]
